Don't copy parent directories of a git flake into the nix store
It's well known that Nix flakes in large repositories can run slowly,
because they copy their state into the store. #3121 and its associated
branch are a fix to this problem.

I am using a Nix flake in large part to produce a reliable dev shell for
developers on the project. My project is O(gigabytes), since it's a game
with many assets.

Each time the dev flake is evaluated, gigabytes are copied into the
store. My flake is not located at the root of the project - instead it's
located in a small subfolder, e.g. `./flake`. Given Nix flake semantics,
my understanding is that this cannot 'reach out' of its directory into
the rest of the repository.

Unfortunately, this still copies gigabytes of data into the store,
because the part of the code that copies the data copies the whole
repository.

This PR is a small optimisation to this process - it looks to see
if the flake ref is pointing to a subdirectory of the repo, and if
so only copies the subdirectory.